refactor: racingCarService 에서 MovingStrategy 의존 제거
