--- src/main/java/racingcar/exception/ErrorResponse.java
@@ -1,14 +1,0 @@
-package racingcar.exception;
-
-public class ErrorResponse {
-
-    private final String message;
-
-    public ErrorResponse(final String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
