+++ src/main/java/racingcar/strategy/RandomMovingStrategy.java
@@ -2,9 +2,7 @@ package racingcar.strategy;
 
 
 import java.util.Random;
-import org.springframework.stereotype.Component;
 
-@Component
 public class RandomMovingStrategy implements MovingStrategy {
 
     private final int MAX_MOVING_NUM = 9;
