+++ src/test/java/racingcar/controller/RacingGameWebControllerTest.java
@@ -1,22 +1,16 @@
 package racingcar.controller;
 
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.core.Is.is;
-
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import racingcar.config.TestConfig;
 import racingcar.dto.GameRequest;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Import({TestConfig.class})
 class RacingGameWebControllerTest {
 
     @LocalServerPort
@@ -28,7 +22,7 @@ class RacingGameWebControllerTest {
     }
 
     @Test
-    @DisplayName("POST 요청이 정상적으로 처리되었는지 확인하고, 반환 값을 확인한다.")
+    @DisplayName("POST 요청이 정상적으로 처리되었는지 확인한다.")
     void createGame() {
         GameRequest gameRequest = new GameRequest("조이,밀리", 5);
 
@@ -37,10 +31,6 @@ class RacingGameWebControllerTest {
                 .body(gameRequest)
                 .when().post("/plays")
                 .then().log().all()
-                .statusCode(HttpStatus.OK.value())
-                .body("winners", is("조이,밀리"))
-                .rootPath("racingCars")
-                .body("[0].name", equalTo("조이"), "[0].position", equalTo(5))
-                .body("[1].name", equalTo("밀리"), "[1].position", equalTo(5));
+                .statusCode(HttpStatus.OK.value());
     }
 }
